latex-service/requirements.txt
flask==3.0.0
gunicorn==21.2.0
requests==2.31.0
python-json-logger==2.0.7
google-cloud-storage==2.10.0
google-cloud-logging==3.8.0
Jinja2==3.1.2
marshmallow==3.20.1
werkzeug==3.0.1
